Refactor tabbed onboarding form to separate component
Make separate import seed form component to allow private key import
page to reuse the same page layout.
