+++ ui/pages/Onboarding/Tabbed/ImportSeed.tsx
@@ -13,6 +13,8 @@ import {
   useAreKeyringsUnlocked,
 } from "../../../hooks"
 import OnboardingRoutes from "./Routes"
+import SharedSeedInput from "../../../components/Shared/SharedSeedInput"
+import ImportForm from "./ImportForm"
 
 type Props = {
   nextPage: string
@@ -85,54 +87,26 @@ export default function ImportSeed(props: Props): ReactElement {
 
   return (
     <>
-      <div className="content fadeIn">
-        <form
-          onSubmit={(event) => {
-            event.preventDefault()
-            importWallet()
-          }}
-        >
-          <div className="portion top">
-            <div className="illustration_import" />
-            <h1 className="serif_header">{t("title")}</h1>
-            <div className="info">{t("subtitle")}</div>
-            <div className="input_wrap">
-              <div
-                id="recovery_phrase"
-                role="textbox"
-                aria-labelledby="recovery_label"
-                tabIndex={0}
-                contentEditable
-                data-empty={recoveryPhrase.length < 1}
-                spellCheck="false"
-                onPaste={(e) => {
-                  e.preventDefault()
-                  const text = e.clipboardData.getData("text/plain").trim()
-                  e.currentTarget.innerText = text
-                  setRecoveryPhrase(text)
-                }}
-                onDrop={(e) => {
-                  e.preventDefault()
-                  const text = e.dataTransfer.getData("text/plain").trim()
-                  e.currentTarget.innerText = text
-                  setRecoveryPhrase(text)
-                }}
-                onInput={(e) => {
-                  setRecoveryPhrase(e.currentTarget.innerText.trim())
-                }}
-              />
-              <div id="recovery_label" className="recovery_label">
-                {t("inputLabel")}
-              </div>
-              {errorMessage && <p className="error">{errorMessage}</p>}
-            </div>
+      <ImportForm
+        title={t("title")}
+        subtitle={t("subtitle")}
+        illustration="doggo_import.svg"
+      >
+        <>
+          <div className="centered">
+            <SharedSeedInput
+              onChange={(seed) => setRecoveryPhrase(seed)}
+              label={t("inputLabel")}
+              errorMessage={errorMessage}
+            />
             {!isEnabled(FeatureFlags.HIDE_IMPORT_DERIVATION_PATH) && (
               <div className="select_wrapper">
                 <OnboardingDerivationPathSelect onChange={setPath} />
               </div>
             )}
           </div>
-          <div className="portion bottom">
+
+          <div className="centered bottom">
             <SharedButton
               style={{ width: "100%", maxWidth: "300px" }}
               size={
@@ -158,31 +132,17 @@ export default function ImportSeed(props: Props): ReactElement {
               </button>
             )}
           </div>
-        </form>
-      </div>
+        </>
+      </ImportForm>
       <style jsx>{`
-        form {
-          all: unset;
-        }
-
-        .content {
+        .centered {
           display: flex;
-          align-items: center;
           flex-direction: column;
+          align-items: center;
           justify-content: space-between;
-        }
-
-        h1 {
-          margin: unset;
-        }
-        .portion {
-          display: flex;
           flex-direction: column;
-          align-items: center;
         }
         .bottom {
-          justify-content: space-between;
-          flex-direction: column;
           margin-top: ${isEnabled(FeatureFlags.HIDE_IMPORT_DERIVATION_PATH)
             ? "48px"
             : "24px"};
@@ -190,24 +150,6 @@ export default function ImportSeed(props: Props): ReactElement {
             ? "24px"
             : "16px"};
         }
-        .illustration_import {
-          background: url("./images/doggo_import.svg");
-          background-size: cover;
-          width: 85px;
-          height: 83px;
-          margin-bottom: 15px;
-        }
-        .serif_header {
-          font-size: 36px;
-          line-height: 42px;
-          margin-bottom: 8px;
-        }
-
-        .info {
-          margin-bottom: 40.5px;
-        }
-
-        .info,
         .help_button {
           width: 320px;
           text-align: center;
@@ -217,80 +159,17 @@ export default function ImportSeed(props: Props): ReactElement {
           font-weight: 500;
         }
         .help_button {
+          width: 320px;
+          text-align: center;
+          font-size: 16px;
+          line-height: 24px;
+          color: var(--green-40);
+          font-weight: 500;
           margin-top: 16px;
-        }
-        .checkbox_wrapper {
-          margin-top: 6px;
-          margin-bottom: 6px;
         }
         .select_wrapper {
           margin-top: ${errorMessage ? "4px" : "15px"};
           width: 320px;
-        }
-        .input_wrap {
-          position: relative;
-        }
-
-        .recovery_label {
-          position: absolute;
-          font-size: 12px;
-          line-height: 16px;
-          transition: all 0.2s ease-in-out;
-          pointer-events: none;
-        }
-
-        #recovery_phrase[data-empty="true"]:not(:focus) ~ .recovery_label {
-          font-size: 16px;
-          line-height: 24px;
-          top: 12px;
-          left: 16px;
-        }
-
-        #recovery_phrase[data-empty="false"] ~ .recovery_label {
-          padding: 0 6px;
-          color: var(--green-40);
-          background: var(--hunter-green);
-          top: -8px;
-          left: 16px;
-        }
-
-        #recovery_phrase {
-          width: 320px;
-          height: 104px;
-          border-radius: 4px;
-          border: 2px solid var(--green-60);
-          padding: 12px 16px;
-          white-space: pre-wrap;
-          word-wrap: break-word;
-          color: var(--white);
-          font-family: inherit;
-          overflow-y: scroll;
-        }
-
-        #recovery_phrase * {
-          word-wrap: break-word;
-          color: var(--white);
-          font-family: inherit;
-        }
-
-        #recovery_phrase:focus ~ .recovery_label {
-          top: -8px;
-          left: 16px;
-          padding: 0 6px;
-          color: var(--trophy-gold);
-          background: var(--hunter-green);
-          transition: all 0.2s ease-in-out;
-          z-index: 999;
-        }
-
-        #recovery_phrase:focus {
-          border: 2px solid var(--trophy-gold);
-          outline: 0;
-          background: var(--hunter-green);
-        }
-
-        .error {
-          color: red;
         }
       `}</style>
     </>
